app_services/requirements.txt
flask
pymongo
python-dotenv
flask_jwt_extended
datetime
requests
bcrypt